feat: rounded crosshair, bigger Ready text, trackball pain slider
- Replace text "+" crosshair with two Rect bars + four Circle end-caps
  to produce rounded arm tips
- Enlarge "Ready" text to 4x font height
- Replace 0-10 keyboard rating with a continuous visual analog slider
  driven by psychopy.hardware.mouse.Mouse (works with HHSC-TRK-2
  trackball or any HID mouse); confirm with click or space
- Add --dev flag to keep the cursor visible for trackball simulation
- Prompt experimenter to "Go to Test" in MMS before opening the
  window and validate the connection before showing instructions;
  show distinct dialogs for socket failure vs MMS error response
- Decode MMS return-code bitfield into named flags
- Switch instruction keys to 4/3 to match mid-task-deterministic

+++ src/medoc/ramp_hold/recorder.py
@@ -23,7 +23,7 @@ class BehaviorRecord:
     hold_onset_s: float | str
     ramp_down_onset_s: float | str
     baseline_return_s: float | str
-    rating: int | str
+    rating: float | str
     rating_rt_ms: float | str
     rating_timeout: int
     trial_end_s: float
